Fix memif re-connection.
This patch request makes memif to behave correctly even if master
or slave drops the connection and possibly re-establishes it again.

Most of this code will be re-written for the single-listener socket
proposal, but for the time being it will make the memif interface easier
to play with by getting rid of the runtime crashes.